@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,7 +24,9 @@ dependencies:
   flutter_screenutil: ^0.7.0
   cupertino_icons: ^0.1.2
   flutter_swiper: ^1.1.6
-  audioplayers: ^0.15.1
+  convex_bottom_bar: ^2.6.0
+  webview_flutter: ^1.0.7
+  shared_preferences: ^0.5.12+4
 
 dev_dependencies:
   flutter_test:
