@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import '../../model/provider/store.dart';
 
 class VideosWidget extends StatefulWidget {
   @override
@@ -15,6 +17,9 @@ class MyComponent extends State<VideosWidget> {
   // props 从 widget.xxx 获取
   @override
   Widget build(BuildContext context) {
-    return Center(child: Text('videos'));
+    return Text(
+      'count：${context.watch<Counter>().count}',
+      key: const Key('counterState'),
+    );
   }
 }
